pyproject.toml
[project]
name = "repl-nix-photomemorylens"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "azure-cognitiveservices-vision-computervision>=0.9.1",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "msrest>=0.7.1",
    "numpy>=2.2.1",
    "opencv-python>=4.10.0.84",
    "pillow>=11.0.0",
    "psycopg2-binary>=2.9.10",
    "routes>=2.5.1",
    "sqlalchemy>=2.0.36",
]
